+++ CHANGELOG.md
@@ -15,6 +15,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [0.10.0] - 2026-09-09
+
+### Changed
+
+- **Breaking:** `object` dependency updated to 0.40
+- **Breaking:** `isr` dependency updated to 0.8.
+
 ## [0.9.0] - 2026-08-27
 
 ### Changed